+++ src/shared/components/ui/tab/stage-tab.stories.tsx
