+++ src/shared/components/ui/tab/stage-tab.tsx
